requirements.txt
# Data collection packages
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

